# contracts/kawsay-estancias-openapi.yaml
...
openapi: 3.0.3
info:
  title: Kawsay Estancias API — contrato propio
  version: 1.2.0
  description: "Contrato propio del dominio de alojamientos de Kawsay Estancias.
    Es la fuente de esquemas para validación AJV y documentación del núcleo
    REST. Conserva nombres de operaciones compatibles con la plantilla
    académica, con modelos y reglas del proyecto: Ecuador, USD, reservas sin
    solapamientos, código de reserva, JWT revocable, idempotencia y outbox. Las
    extensiones administrativas se documentan adicionalmente mediante DTO de
    NestJS. No hay cobros reales ni entrega automática de webhooks."
  x-origin: Diseñado para Kawsay Estancias; la plantilla se conserva por separado
    como referencia académica.
servers:
  - url: /api/v1
    description: API del despliegue desde el que se abre el contrato
  - url: https://booking-prototipo-alojamientos.vercel.app/api/v1
    description: Producción actual de Kawsay Estancias
security:
  - bearer: []
  - cookie: []
...
paths:
...
  /details/changes:
    post:
      operationId: post_details_changes
      tags:
        - Alojamientos
      summary: Consultar cambios de catálogo para sincronizar otro sistema
      description: ""
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/DetailsChangesRequest"
      responses:
        "200":
          description: Operación completada.
          content:
            application/json:
              schema:
                type: object
                required:
                  - request_id
                  - data
                  - next_page
                properties:
                  request_id:
                    type: string
                    format: uuid
                  data:
                    type: object
                    properties:
                      from:
                        type: string
                        format: date-time
                      next:
                        type: string
                        format: date-time
                      total_changes:
                        type: integer
                      changes:
                        type: object
                        properties:
                          accommodations:
                            type: array
                            items:
                              type: object
                              properties:
                                id:
                                  type: integer
                                  minimum: 1
                                updated_at:
                                  type: string
                                  format: date-time
                                published:
                                  type: boolean
                  next_page:
                    type: string
                    nullable: true
        "400":
          $ref: "#/components/responses/ProblemDetails400"
        "401":
          $ref: "#/components/responses/ProblemDetails401"
        "403":
          $ref: "#/components/responses/ProblemDetails403"
        "409":
          $ref: "#/components/responses/ProblemDetails409"
...
  /reviews/scores:
    post:
      operationId: post_reviews_scores
      tags:
        - Alojamientos
      summary: Consultar puntuaciones de alojamientos
      description: ""
      security: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/ReviewsScoresRequest"
      responses:
        "200":
          description: Operación completada.
          content:
            application/json:
              schema:
                type: object
                required:
                  - request_id
                  - data
                  - next_page
                properties:
                  request_id:
                    type: string
                    format: uuid
                  data:
                    type: array
                    items:
                      type: object
                      properties: {}
                  next_page:
                    type: string
                    nullable: true
        "400":
          $ref: "#/components/responses/ProblemDetails400"
        "409":
          $ref: "#/components/responses/ProblemDetails409"
  /orders/preview:
    post:
      operationId: post_orders_preview
      tags:
        - Reservas
      summary: Cotizar una reserva antes de confirmarla
      description: Solo viajeros. La cotización pertenece al usuario y vence a los 10
        minutos.
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/OrderPreviewRequest"
      responses:
        "200":
          description: Operación completada.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/QuoteResponse"
        "400":
          $ref: "#/components/responses/ProblemDetails400"
        "401":
          $ref: "#/components/responses/ProblemDetails401"
        "403":
          $ref: "#/components/responses/ProblemDetails403"
        "409":
          $ref: "#/components/responses/ProblemDetails409"
  /orders/create:
    post:
      operationId: post_orders_create
      tags:
        - Reservas
      summary: Confirmar reserva y obtener su código
      description: Revalida las noches dentro de la transacción. Guarda reserva,
        factura, respuesta idempotente y evento ORDER_CONFIRMED. El campo
        locator es el código de reserva. El pago es simulado.
      parameters:
        - name: Idempotency-Key
          in: header
          required: true
          schema:
            type: string
            format: uuid
          description: UUID v4. Repetir la misma clave y cuerpo devuelve la misma reserva;
            cambiar el cuerpo produce 409.
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/OrderCreateRequest"
      responses:
        "201":
          description: Operación completada.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Reservation"
        "400":
          $ref: "#/components/responses/ProblemDetails400"
        "401":
          $ref: "#/components/responses/ProblemDetails401"
        "403":
          $ref: "#/components/responses/ProblemDetails403"
        "409":
          $ref: "#/components/responses/ProblemDetails409"
  /orders:
    get:
      operationId: get_orders
      tags:
        - Reservas
      summary: Consultar Mis reservas
      description: ""
      responses:
        "200":
          description: Operación completada.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ReservationList"
        "400":
          $ref: "#/components/responses/ProblemDetails400"
        "401":
          $ref: "#/components/responses/ProblemDetails401"
        "403":
          $ref: "#/components/responses/ProblemDetails403"
        "409":
          $ref: "#/components/responses/ProblemDetails409"
  /orders/{orderId}:
    get:
      operationId: get_orders__orderId_
      tags:
        - Reservas
      summary: Consultar una reserva propia o con rol administrador
      description: ""
      parameters:
        - name: orderId
          in: path
          required: true
          schema:
            type: string
            format: uuid
      responses:
        "200":
          description: Operación completada.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Reservation"
        "400":
          $ref: "#/components/responses/ProblemDetails400"
        "401":
          $ref: "#/components/responses/ProblemDetails401"
        "403":
          $ref: "#/components/responses/ProblemDetails403"
        "409":
          $ref: "#/components/responses/ProblemDetails409"
  /orders/{orderId}/modify:
    post:
      operationId: post_orders__orderId__modify
      tags:
        - Reservas
      summary: Modificar fechas sin solaparse con otra reserva
      description: ""
      parameters:
        - name: orderId
          in: path
          required: true
          schema:
            type: string
            format: uuid
        - name: Idempotency-Key
          in: header
          required: true
          schema:
            type: string
            format: uuid
          description: UUID v4. Repetir la misma clave y cuerpo devuelve la misma reserva;
            cambiar el cuerpo produce 409.
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/OrderModifyRequest"
      responses:
        "200":
          description: Operación completada.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Reservation"
        "400":
          $ref: "#/components/responses/ProblemDetails400"
        "401":
          $ref: "#/components/responses/ProblemDetails401"
        "403":
          $ref: "#/components/responses/ProblemDetails403"
        "409":
          $ref: "#/components/responses/ProblemDetails409"
  /orders/{orderId}/cancel:
    post:
      operationId: post_orders__orderId__cancel
      tags:
        - Reservas
      summary: Cancelar y liberar las noches reservadas
      description: ""
      parameters:
        - name: orderId
          in: path
          required: true
          schema:
            type: string
            format: uuid
        - name: Idempotency-Key
          in: header
          required: true
          schema:
            type: string
            format: uuid
          description: UUID v4. Repetir la misma clave y cuerpo devuelve la misma reserva;
            cambiar el cuerpo produce 409.
      responses:
        "200":
          description: Operación completada.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Reservation"
        "400":
          $ref: "#/components/responses/ProblemDetails400"
        "401":
          $ref: "#/components/responses/ProblemDetails401"
        "403":
          $ref: "#/components/responses/ProblemDetails403"
        "409":
          $ref: "#/components/responses/ProblemDetails409"
...
  /auth/me:
    get:
      operationId: get_auth_me
      tags:
        - Identidad
      summary: Consultar la identidad autenticada
      description: ""
      responses:
        "200":
          description: Operación completada.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/User"
        "400":
          $ref: "#/components/responses/ProblemDetails400"
        "401":
          $ref: "#/components/responses/ProblemDetails401"
        "403":
          $ref: "#/components/responses/ProblemDetails403"
        "409":
          $ref: "#/components/responses/ProblemDetails409"
...
  /catalog:
    get:
      operationId: get_catalog
      tags:
        - Alojamientos
      summary: Consultar catálogo público
      description: ""
      security: []
      parameters:
        - name: nombre
          in: query
          schema:
            type: string
      responses:
        "200":
          description: Operación completada.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/AccommodationList"
        "400":
          $ref: "#/components/responses/ProblemDetails400"
        "409":
          $ref: "#/components/responses/ProblemDetails409"
  /catalog/{id}:
    get:
      operationId: get_catalog__id_
      tags:
        - Alojamientos
      summary: Consultar un alojamiento publicado
      description: ""
      security: []
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
            minimum: 1
      responses:
        "200":
          description: Operación completada.
          content:
            application/json:
              schema:
                type: object
                required:
                  - request_id
                  - data
                  - next_page
                properties:
                  request_id:
                    type: string
                    format: uuid
                  data:
                    $ref: "#/components/schemas/Accommodation"
                  next_page:
                    type: string
                    nullable: true
        "400":
          $ref: "#/components/responses/ProblemDetails400"
        "409":
          $ref: "#/components/responses/ProblemDetails409"
  /cities:
    get:
      operationId: get_cities
      tags:
        - Alojamientos
      summary: Consultar destinos de Ecuador
      description: ""
      security: []
      responses:
        "200":
          description: Operación completada.
          content:
            application/json:
              schema:
                type: object
                required:
                  - request_id
                  - data
                  - next_page
                properties:
                  request_id:
                    type: string
                    format: uuid
                  data:
                    type: array
                    items:
                      type: object
                      properties:
                        id:
                          type: integer
                          minimum: 1
                        name:
                          type: string
                        country:
                          type: string
                  next_page:
                    type: string
                    nullable: true
        "400":
          $ref: "#/components/responses/ProblemDetails400"
        "409":
          $ref: "#/components/responses/ProblemDetails409"
...
  /webhooks:
    get:
      operationId: get_webhooks
      tags:
        - Integración futura
      summary: Consultar suscripciones sin revelar secretos
      description: ""
      responses:
        "200":
          description: Operación completada.
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/WebhookView"
        "400":
          $ref: "#/components/responses/ProblemDetails400"
        "401":
          $ref: "#/components/responses/ProblemDetails401"
        "403":
          $ref: "#/components/responses/ProblemDetails403"
        "409":
          $ref: "#/components/responses/ProblemDetails409"
    post:
      operationId: post_webhooks
      tags:
        - Integración futura
      summary: Registrar un destino para futura entrega de eventos
      description: Solo administrador. Guarda una suscripción HTTPS y su secreto. Esta
        versión no envía solicitudes al destino registrado.
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/WebhookSubscription"
      responses:
        "201":
          description: Operación completada.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/WebhookView"
        "400":
          $ref: "#/components/responses/ProblemDetails400"
        "401":
          $ref: "#/components/responses/ProblemDetails401"
        "403":
          $ref: "#/components/responses/ProblemDetails403"
        "409":
          $ref: "#/components/responses/ProblemDetails409"
  /webhooks/{id}:
    delete:
      operationId: delete_webhooks__id_
      tags:
        - Integración futura
      summary: Eliminar una suscripción
      description: ""
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
            format: uuid
      responses:
        "204":
          description: Operación completada sin cuerpo.
        "400":
          $ref: "#/components/responses/ProblemDetails400"
        "401":
          $ref: "#/components/responses/ProblemDetails401"
        "403":
          $ref: "#/components/responses/ProblemDetails403"
        "409":
          $ref: "#/components/responses/ProblemDetails409"
components:
  securitySchemes:
    bearer:
      type: http
      scheme: bearer
      bearerFormat: JWT
    cookie:
      type: apiKey
      in: cookie
      name: booking_session
  schemas:
...
    DetailsChangesRequest:
      type: object
      required:
        - last_change
      properties:
        last_change:
          type: string
          format: date-time
        filters:
          type: object
          properties:
            countries:
              type: array
              items:
                type: string
            cities:
              type: array
              items:
                type: integer
                minimum: 1
...
    ReviewsScoresRequest:
      type: object
      required:
        - accommodations
      properties:
        accommodations:
          type: array
          items:
            type: integer
            minimum: 1
        languages:
          type: array
          items:
            type: string
    OrderPreviewRequest:
      type: object
      required:
        - accommodation_id
        - product_id
        - guests
      properties:
        accommodation_id:
          type: integer
          minimum: 1
        product_id:
          type: string
          format: uuid
        guests:
          $ref: "#/components/schemas/AccommodationsGuests"
...
    OrderCreateRequest:
      type: object
      required:
        - order_preview_id
        - payment_reference
        - customer_details
      properties:
        order_preview_id:
          type: string
          format: uuid
        payment_reference:
          type: string
          pattern: ^DEMO-
          description: Referencia de pago simulado; el prototipo no cobra dinero.
        customer_details:
          $ref: "#/components/schemas/CustomerDetails"
    OrderModifyRequest:
      type: object
      properties:
        checkin:
          type: string
          format: date
        checkout:
          type: string
          format: date
        guests:
          $ref: "#/components/schemas/AccommodationsGuests"
...
    User:
      type: object
      required:
        - id
        - name
        - email
        - role
      properties:
        id:
          type: string
          format: uuid
        name:
          type: string
        email:
          type: string
          format: email
        role:
          type: string
          enum:
            - admin
            - customer
...
    Reservation:
      type: object
      required:
        - order_id
        - locator
        - status
        - accommodation_details
        - checkin
        - checkout
        - guests
        - total_price
        - currency
        - creation_date
      properties:
        order_id:
          type: string
          format: uuid
        locator:
          type: string
          pattern: ^BP-[A-F0-9]{8}$
          description: Código de reserva visible en la confirmación y Mis reservas.
          example: BP-41A92454
        status:
          type: string
          enum:
            - CONFIRMED
            - CANCELLED
        accommodation_details:
          type: object
          required:
            - id
          properties:
            id:
              type: integer
              minimum: 1
        checkin:
          type: string
          format: date
        checkout:
          type: string
          format: date
        guests:
          $ref: "#/components/schemas/AccommodationsGuests"
        customer_details:
          $ref: "#/components/schemas/CustomerDetails"
        total_price:
          type: number
          minimum: 0
        currency:
          type: string
          enum:
            - USD
          default: USD
        creation_date:
          type: string
          format: date-time
        estancia_estado:
          type: string
          enum:
            - CONFIRMADA
            - EN_CURSO
            - COMPLETADA
            - CANCELADA
        puede_resenar:
          type: boolean
        _links:
          type: object
          properties:
            self:
              type: string
            modify:
              type: string
            cancel:
              type: string
...
    AccommodationList:
      type: object
      required:
        - request_id
        - data
        - next_page
      properties:
        request_id:
          type: string
          format: uuid
        data:
          type: array
          items:
            $ref: "#/components/schemas/Accommodation"
        next_page:
          type: string
          nullable: true
...
    QuoteResponse:
      type: object
      required:
        - request_id
        - data
        - next_page
      properties:
        request_id:
          type: string
          format: uuid
        data:
          $ref: "#/components/schemas/Quote"
        next_page:
          type: string
          nullable: true
    ReservationList:
      type: object
      required:
        - request_id
        - data
        - next_page
      properties:
        request_id:
          type: string
          format: uuid
        data:
          type: array
          items:
            $ref: "#/components/schemas/Reservation"
        next_page:
          type: string
          nullable: true
...
  responses:
    ProblemDetails400:
      description: Solicitud inválida según el contrato o las reglas del dominio.
      content:
        application/problem+json:
          schema:
            $ref: "#/components/schemas/Problem"
    ProblemDetails401:
      description: Sesión ausente o vencida.
      content:
        application/problem+json:
          schema:
            $ref: "#/components/schemas/Problem"
    ProblemDetails403:
      description: Permisos insuficientes.
      content:
        application/problem+json:
          schema:
            $ref: "#/components/schemas/Problem"
...
    ProblemDetails409:
      description: Fechas ocupadas, cotización vencida o clave idempotente reutilizada
        con otro contenido.
      content:
        application/problem+json:
          schema:
            $ref: "#/components/schemas/Problem"
...